refactoring: miranum-spring-security
