+++ examples/inquiry-integration-service/pom.xml
@@ -37,7 +37,7 @@
         <!-- miranum libs -->
         <dependency>
             <groupId>io.miragon.miranum</groupId>
-            <artifactId>spring-security-starter</artifactId>
+            <artifactId>miranum-spring-security</artifactId>
             <version>${project.version}</version>
         </dependency>
         <dependency>
