+++ examples/inquiry-integration-service/src/main/java/io/miragon/miranum/inquiry/application/CapacityCheckedUseCase.java
@@ -1,12 +1,12 @@
 package io.miragon.miranum.inquiry.application;
 
+import io.miragon.miranum.auth.api.UserAuthenticationProvider;
 import io.miragon.miranum.connect.task.api.TaskApi;
 import io.miragon.miranum.connect.task.api.command.CompleteTaskCommand;
 import io.miragon.miranum.inquiry.application.port.in.CapacityChecked;
 import io.miragon.miranum.inquiry.application.port.in.model.CapacityCheckedCommand;
 import io.miragon.miranum.inquiry.application.port.out.InquiryRepository;
 import io.miragon.miranum.inquiry.domain.Inquiry;
-import io.miragon.miranum.platform.security.authentication.UserAuthenticationProvider;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
