+++ examples/single-deployment-unit-example/pom.xml
@@ -84,7 +84,7 @@
         <!-- miranum libs -->
         <dependency>
             <groupId>io.miragon.miranum</groupId>
-            <artifactId>spring-security-starter</artifactId>
+            <artifactId>miranum-spring-security</artifactId>
             <version>${project.version}</version>
         </dependency>
         <dependency>
