+++ libs/auth-core/src/main/java/io/miragon/miranum/auth/api/UserAuthenticationProvider.java
@@ -1,0 +1,10 @@
+package io.miragon.miranum.auth.api;
+
+import java.util.List;
+
+public interface UserAuthenticationProvider {
+
+    String getLoggedInUser();
+
+    List<String> getLoggedInUserRoles();
+}
