+++ libs/spring-security/src/main/java/io/miragon/miranum/security/authentication/UserAuthenticationProviderImpl.java
@@ -1,0 +1,46 @@
+package io.miragon.miranum.security.authentication;
+
+import io.miragon.miranum.auth.api.UserAuthenticationProvider;
+import io.miragon.miranum.security.SpringSecurityProperties;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.context.annotation.Profile;
+import org.springframework.lang.NonNull;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.oauth2.jwt.Jwt;
+import org.springframework.stereotype.Component;
+
+import java.util.List;
+
+/**
+ * User authentication provider.
+ * Extracts the username from the token.
+ */
+@Component
+@Profile("!no-security")
+@RequiredArgsConstructor
+@Slf4j
+public class UserAuthenticationProviderImpl implements UserAuthenticationProvider {
+
+    public static final String NAME_UNAUTHENTICATED_USER = "unauthenticated";
+    private final SpringSecurityProperties springSecurityProperties;
+
+    @Override
+    @NonNull
+    public String getLoggedInUser() {
+        final Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        if (authentication.getPrincipal() instanceof Jwt jwt) {
+            return (String) jwt.getClaims().get(springSecurityProperties.getServer().getUserNameAttribute());
+        }
+        return NAME_UNAUTHENTICATED_USER;
+    }
+
+    @Override
+    @NonNull
+    public List<String> getLoggedInUserRoles() {
+        final Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        return authentication.getAuthorities().stream().map(Object::toString).map(obj -> obj.replaceFirst("^ROLE_", "")).toList();
+    }
+
+}
