+++ libs/spring-security/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,0 +1,1 @@
+io.miragon.miranum.security.autoconfiguration.SpringSecurityAutoConfiguration
