+++ libs/spring-security/src/main/resources/application-spring-security-starter.yaml
