+++ platform/engine/core/pom.xml
@@ -49,6 +49,13 @@
             <artifactId>jackson-datatype-jsr310</artifactId>
         </dependency>
 
+        <!-- Miranum -->
+        <dependency>
+            <groupId>io.miragon.miranum</groupId>
+            <artifactId>miranum-auth-core</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+
         <!-- Miranum Connect -->
         <dependency>
             <groupId>io.miragon.miranum</groupId>
