--- platform/engine/core/src/main/java/io/miragon/miranum/platform/engine/api/AppAuthenticationProvider.java
@@ -1,9 +1,0 @@
-package io.miragon.miranum.platform.engine.api;
-
-import java.util.List;
-
-public interface AppAuthenticationProvider {
-    String getCurrentUserId();
-
-    List<String> getCurrentUserGroups();
-}
