+++ platform/engine/task/pom.xml
@@ -16,6 +16,10 @@
             <groupId>org.springframework.boot</groupId>
             <artifactId>spring-boot-starter-web</artifactId>
         </dependency>
+        <dependency>
+            <groupId>org.springframework.boot</groupId>
+            <artifactId>spring-boot-starter-data-jpa</artifactId>
+        </dependency>
 
         <dependency>
             <groupId>org.springdoc</groupId>
@@ -36,18 +40,15 @@
         <!-- Miranum -->
         <dependency>
             <groupId>io.miragon.miranum</groupId>
-            <artifactId>miranum-engine-user</artifactId>
-            <version>${project.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>io.miragon.miranum</groupId>
             <artifactId>miranum-engine-core</artifactId>
             <version>${project.version}</version>
+            <scope>provided</scope>
         </dependency>
         <dependency>
             <groupId>io.miragon.miranum</groupId>
-            <artifactId>spring-security-starter</artifactId>
+            <artifactId>miranum-auth-core</artifactId>
             <version>${project.version}</version>
+            <scope>provided</scope>
         </dependency>
 
         <!-- miranum connect -->
