--- platform/libs/pom.xml
@@ -1,20 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns="http://maven.apache.org/POM/4.0.0"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-    <artifactId>miranum-platform-libs</artifactId>
-    <name>miranum-platform-libs</name>
-    <packaging>pom</packaging>
-
-    <parent>
-        <groupId>io.miragon.miranum</groupId>
-        <artifactId>miranum-platform</artifactId>
-        <version>${revision}</version>
-    </parent>
-
-    <modules>
-        <module>spring-security</module>
-    </modules>
-
-
-</project>
