--- platform/libs/spring-security/spring-security-core/src/main/java/io/miragon/miranum/platform/security/authentication/UserAuthenticationProvider.java
@@ -1,13 +1,0 @@
-package io.miragon.miranum.platform.security.authentication;
-
-import org.springframework.lang.NonNull;
-
-import java.util.List;
-
-public interface UserAuthenticationProvider {
-
-    @NonNull
-    String getLoggedInUser();
-
-    List<String> getLoggedInUserRoles();
-}
