--- platform/libs/spring-security/spring-security-starter/src/main/java/io/miragon/miranum/platform/security/autoconfiguration/SpringSecurityAutoConfiguration.java
@@ -1,16 +1,0 @@
-package io.miragon.miranum.platform.security.autoconfiguration;
-
-
-import io.miragon.miranum.platform.security.SpringSecurityProperties;
-import org.springframework.boot.context.properties.EnableConfigurationProperties;
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.PropertySource;
-
-/**
- * Auto-configuration used to configure Spring security.
- */
-@EnableConfigurationProperties(SpringSecurityProperties.class)
-@ComponentScan(basePackages = "io.miragon.miranum.platform.security")
-@PropertySource(value = {"classpath:application-spring-security-starter.yaml"}, factory = YamlPropertySourceFactory.class)
-public class SpringSecurityAutoConfiguration {
-}
