--- platform/libs/spring-security/spring-security-starter/src/main/java/io/miragon/miranum/platform/security/autoconfiguration/YamlPropertySourceFactory.java
@@ -1,24 +1,0 @@
-package io.miragon.miranum.platform.security.autoconfiguration;
-
-import org.springframework.beans.factory.config.YamlPropertiesFactoryBean;
-import org.springframework.core.env.PropertiesPropertySource;
-import org.springframework.core.env.PropertySource;
-import org.springframework.core.io.support.EncodedResource;
-import org.springframework.core.io.support.PropertySourceFactory;
-
-import java.io.IOException;
-import java.util.Objects;
-import java.util.Properties;
-
-public class YamlPropertySourceFactory implements PropertySourceFactory {
-    @Override
-    public PropertySource<?> createPropertySource(String name, EncodedResource resource) throws IOException {
-        final YamlPropertiesFactoryBean factory = new YamlPropertiesFactoryBean();
-        factory.setResources(resource.getResource());
-        final Properties properties = factory.getObject();
-        return new PropertiesPropertySource(
-                Objects.requireNonNull(resource.getResource().getFilename()),
-                Objects.requireNonNull(properties)
-        );
-    }
-}
