--- platform/libs/spring-security/spring-security-starter/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,1 +1,0 @@
-io.miragon.miranum.platform.security.autoconfiguration.SpringSecurityAutoConfiguration
