+++ platform/pom.xml
@@ -16,7 +16,6 @@
     <modules>
         <module>deployment</module>
         <module>engine</module>
-        <module>libs</module>
     </modules>
 
     <!-- Global dependencies for all platform modules -->
